requirements.txt
Flask>=2.3.0
pymongo>=4.5.0
python-dotenv>=1.0.0
bson>=0.5.10

